Eliminate warnings.
git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@1367 d4fa192b-c00b-0410-8231-f00ffab90ce4

+++ lib/libvcl/vcc_priv.h
@@ -43,4 +43,4 @@ extern const char *vcl_tnames[256];
 void vcl_init_tnames(void);
 void vcl_output_lang_h(struct vsb *sb);
 
-#define PF(t)	((t)->e - (t)->b), (t)->b
+#define PF(t)	(int)((t)->e - (t)->b), (t)->b
